fix(xtoken): unblock projection-build self-test and tensorboard logging
tools/x_token/minimal_projection_via_multitoken.py: add missing `import re`
(extract_keywords at line 286 calls `re.sub` but the module was never
imported, so the special-token-mapping path crashed with NameError); force
the post-save self-test dummy and sparse-CSR values to float32 (bf16 +
sparse-CSR matmul is unimplemented in PyTorch's MKL backend and crashed the
script after the .pt file was already saved).

examples/configs/xtoken_off_policy_distillation.yaml: add the missing
`logger.tensorboard` subsection. `Logger.__init__` looks up
`cfg["tensorboard"]` whenever `tensorboard_enabled=true`, so the YAML's
omission triggered a KeyError as soon as a user opted into TB logging.

+++ tools/x_token/minimal_projection_via_multitoken.py
@@ -14,6 +14,7 @@
 import argparse
 import difflib
 import os
+import re
 from collections import defaultdict
 
 import torch
@@ -922,13 +923,13 @@ if __name__ == "__main__":
 
     # Create a dummy likelihood tensor: [BATCH, SEQ, source_vocab_size]
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-    dummy_tensor = torch.randn(1, 4096, tokenizer_student_total_vocab_size, device=device, dtype=torch.bfloat16)
+    dummy_tensor = torch.randn(1, 4096, tokenizer_student_total_vocab_size, device=device, dtype=torch.float32)
 
     # Transform this tensor using the projection map
     projected_tensor = project_token_likelihoods(
         dummy_tensor,
         top_k_indices.to(device),
-        top_k_likelihoods.to(device),
+        top_k_likelihoods.float().to(device),
         tokenizer_teacher_total_vocab_size,
         device
     )
